Refactor ASN1 and ASN1Exception into standalone classes in preparation of KDF rework.
